remove old selling info controller
